deadline_/requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
decorator==4.0.11
Django==1.10.5
django-celery-results==1.0.1
django-cors-headers==2.0.2
django-cors-middleware==1.3.1
djangorestframework==3.5.4
fake-factory==9999.9.9
ipdb==0.10.2
ipython==5.2.2
ipython-genutils==0.1.0
kombu==4.0.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
PyJWT==1.4.2
pytz==2016.10
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
